requirements.txt
fitz==0.0.1.dev2
langchain==0.3.20
langchain_chroma==0.2.2
langchain_community==0.3.19
langchain_core==0.3.45
langchain_groq==0.2.5
langchain_huggingface==0.1.2
langchain_text_splitters==0.3.6
Pillow==11.1.0
python-dotenv==1.0.1
streamlit==1.40.2
